Reorganising file composition
