--- Codigo/ControlTemperatura/modo.h
@@ -1,1 +1,0 @@
-typedef enum{FRIO,CALOR,CALIBRACION}modo_t;
