+++ Codigo/firmware/dev/ConfiguracionADC.h
