+++ Codigo/firmware/dev/ConfiguracionSerial.c
