+++ Codigo/firmware/dev/funciones.h
