+++ Documentacion/Mediciones/openloopHand1A.csv
